+++ android/src/main/kotlin/net/zonble/flutter_platform_alert/FlutterPlatformAlertPlugin.kt
@@ -3,6 +3,7 @@ package net.zonble.flutter_platform_alert
 import android.app.Activity
 import android.app.AlertDialog
 import android.content.Context
+import android.content.res.Configuration
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
 import android.graphics.drawable.BitmapDrawable
@@ -10,7 +11,6 @@ import android.graphics.drawable.Drawable
 import android.media.RingtoneManager
 import android.os.Build
 import android.util.Base64
-import androidx.annotation.NonNull
 import androidx.annotation.RequiresApi
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.activity.ActivityAware
@@ -27,7 +27,7 @@ class FlutterPlatformAlertPlugin : FlutterPlugin, MethodCallHandler, ActivityAwa
 
   @Suppress("UNCHECKED_CAST")
   @RequiresApi(Build.VERSION_CODES.N)
-  override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: Result): Unit =
+  override fun onMethodCall(call: MethodCall, result: Result): Unit =
     when (call.method) {
       "playAlertSound" -> {
         val notification = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION)
@@ -47,7 +47,7 @@ class FlutterPlatformAlertPlugin : FlutterPlugin, MethodCallHandler, ActivityAwa
 
           AlertDialog.Builder(
             this.activity,
-            R.style.AlertDialogCustom
+            getDialogStyle()
           ).setTitle(windowTitle).setMessage(text).apply {
             when (alertStyle) {
               "abortRetryIgnore" ->
@@ -56,7 +56,7 @@ class FlutterPlatformAlertPlugin : FlutterPlugin, MethodCallHandler, ActivityAwa
                   .setNegativeButton(R.string.abort) { _, _ -> result.success("abort") }
               "cancelTryContinue" ->
                 setPositiveButton(R.string.try_again) { _, _ -> result.success("try_again") }
-                  .setNeutralButton(R.string.continue_button) { _, _ -> result.success("continue") }
+                  .setNeutralButton(R.string.continue_button) { _, _ -> result.success("continue" ) }
                   .setNegativeButton(R.string.cancel) { _, _ -> result.success("cancel") }
               "okCancel" ->
                 setPositiveButton(R.string.ok) { _, _ -> result.success("ok") }
@@ -90,7 +90,7 @@ class FlutterPlatformAlertPlugin : FlutterPlugin, MethodCallHandler, ActivityAwa
 
           val builder = AlertDialog.Builder(
             this.activity,
-            R.style.AlertDialogCustom
+            getDialogStyle()
           ).setTitle(windowTitle).setMessage(text)
           var buttonCount = 0
           if (positiveButtonTitle.isNotEmpty()) {
@@ -105,7 +105,7 @@ class FlutterPlatformAlertPlugin : FlutterPlugin, MethodCallHandler, ActivityAwa
             builder.setNeutralButton(neutralButtonTitle) { _, _ -> result.success("neutral_button") }
             buttonCount += 1
           }
-          if (buttonCount==0) {
+          if (buttonCount == 0) {
             builder.setPositiveButton("OK") { _, _ -> result.success("other") }
             buttonCount += 1
           }
@@ -123,14 +123,25 @@ class FlutterPlatformAlertPlugin : FlutterPlugin, MethodCallHandler, ActivityAwa
       else -> result.notImplemented()
     }
 
-  override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
+  @RequiresApi(Build.VERSION_CODES.LOLLIPOP_MR1)
+  private fun getDialogStyle(): Int {
+    if (this.context != null) {
+      val nightModeFlags: Int = this.context!!.resources.configuration.uiMode and
+              Configuration.UI_MODE_NIGHT_MASK
+      when (nightModeFlags) {
+        Configuration.UI_MODE_NIGHT_NO -> return android.R.style.Theme_DeviceDefault_Light_Dialog_Alert
+      }
+    }
+    return android.R.style.Theme_DeviceDefault_Dialog_Alert
+  }
+
+  override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
     channel = MethodChannel(flutterPluginBinding.binaryMessenger, "flutter_platform_alert")
     channel.setMethodCallHandler(this)
     context = flutterPluginBinding.applicationContext
   }
 
-
-  override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
+  override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {
     channel.setMethodCallHandler(null)
     context = null
   }