+++ ios/Classes/SwiftFlutterPlatformAlertPlugin.swift
@@ -129,7 +129,7 @@ public class SwiftFlutterPlatformAlertPlugin: NSObject, FlutterPlugin {
             let alertStyleString = args["alertStyle"] as? String ?? ""
             let alertStyle = FlutterPlatformAlertStyle(rawValue: alertStyleString) ?? FlutterPlatformAlertStyle.ok
             let buttons = alertStyle.buttons
-            let preferredStyle: UIAlertController.Style  = .alert
+            let preferredStyle: UIAlertController.Style  = getAlertStyle(args["iosAlertStyle"] as? String)
             let controller = UIAlertController(title: windowTitle, message: text, preferredStyle: preferredStyle)
             for i in 0..<buttons.count {
                 let button = buttons[i]
@@ -157,26 +157,26 @@ public class SwiftFlutterPlatformAlertPlugin: NSObject, FlutterPlugin {
             var actions = [UIAlertAction]()
             if let positiveButton = args["positiveButtonTitle"] as? String,
                positiveButton.isEmpty == false {
-                let buttonStyle = style(forButtonTitle: positiveButton)
+                let buttonStyle = getButtonStyle(args["positiveButtonStyle"] as? String)
                 actions.append(UIAlertAction(title: positiveButton, style: buttonStyle) { action in
                     result(CustomAlertButton.positiveButton.rawValue)
                 })
             }
             if let neutralButton = args["neutralButtonTitle"] as? String,
                neutralButton.isEmpty == false {
-                let buttonStyle = style(forButtonTitle: neutralButton)
+                let buttonStyle = getButtonStyle(args["neutralButtonStyle"] as? String)
                 actions.append(UIAlertAction(title: neutralButton, style: buttonStyle) { action in
                     result(CustomAlertButton.neutralButton.rawValue)
                 })
             }
             if let negativeButton = args["negativeButtonTitle"] as? String,
                negativeButton.isEmpty == false {
-                let buttonStyle = style(forButtonTitle: negativeButton)
+                let buttonStyle = getButtonStyle(args["negativeButtonStyle"] as? String)
                 actions.append(UIAlertAction(title: negativeButton, style: buttonStyle) { action in
                     result(CustomAlertButton.negativeButton.rawValue)
                 })
             }
-            let preferredStyle: UIAlertController.Style  = .alert
+            let preferredStyle: UIAlertController.Style  = getAlertStyle(args["iosAlertStyle"] as? String)
             let controller = UIAlertController(title: windowTitle, message: text, preferredStyle: preferredStyle)
             for action in actions {
                 controller.addAction(action)
@@ -187,5 +187,25 @@ public class SwiftFlutterPlatformAlertPlugin: NSObject, FlutterPlugin {
             result(FlutterMethodNotImplemented)
         }
 
+    }
+
+    private func getButtonStyle(_ style: String?) -> UIAlertAction.Style {
+        switch style {
+        case "cancel":
+            return .cancel
+        case "destructive":
+            return .destructive
+        default:
+            return .default
+        }
+    }
+
+    private func getAlertStyle(_ style: String?) -> UIAlertController.Style {
+        switch style {
+        case "actionSheet":
+            return .actionSheet
+        default:
+            return .alert
+        }
     }
 }