+++ lib/src/styles/ios_alert_style.dart
@@ -1,0 +1,10 @@
+/// Enum mapping for the [UIAlertController.Style](https://developer.apple.com/documentation/uikit/uialertcontroller/style).
+enum IosAlertStyle {
+  /// Alert dialog.
+  /// Native equivalent to [CupertinoAlertDialog](https://api.flutter.dev/flutter/cupertino/CupertinoAlertDialog-class.html).
+  alertDialog,
+
+  /// Bottom action sheet.
+  /// Native equivalent to [CupertinoActionSheet](https://api.flutter.dev/flutter/cupertino/CupertinoActionSheet-class.html).
+  actionSheet,
+}