+++ lib/src/styles/ios_button_style.dart
@@ -1,0 +1,11 @@
+/// Enum mapping for the [UIAlertAction.Style](https://developer.apple.com/documentation/uikit/uialertaction/style).
+enum IosButtonStyle {
+  /// Bold label that represents to cancel something.
+  cancel,
+
+  /// Red label that represents to change or delete something.
+  destructive,
+
+  /// Default label style.
+  normal,
+}