+++ test/flutter_platform_alert_test.dart
@@ -1,7 +1,3 @@
-import 'package:flutter/services.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_platform_alert/flutter_platform_alert.dart';
-
 void main() {
   // const MethodChannel channel = MethodChannel('flutter_platform_alert');
   //